requirements.txt
APScheduler==3.6.3
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
decorator==4.4.2
dialogflow==1.1.0
feedparser==6.0.2
Flask==1.1.2
fuzzywuzzy==0.18.0
gnewsclient==1.12
google-api-core==1.26.2
google-auth==1.28.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-telegram-bot==12.8
pytz==2021.1
requests==2.25.1
rsa==4.7.2
sgmllib3k==1.0.0
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
Werkzeug==1.0.1
